refatorando as actions e o reducer do area
