+++ src/pages/FoodsOrigin.jsx
@@ -17,7 +17,6 @@ class ExploreArea extends Component {
 
     this.state = {
       data: [],
-      loading: true,
     };
   }
 
@@ -32,6 +31,7 @@ class ExploreArea extends Component {
     if (meals.length === numberToComper) {
       return <Redirect to={ `/comidas/${meals[0].idMeal}` } />;
     }
+    if (isFetching) return <p>Loading</p>;
     if (!meals.length && !isFetching) return this.renderAlertError();
     if (meals.length === 1) return this.redirectToRecipeDetail();
     return this.renderRecipes();
@@ -39,7 +39,7 @@ class ExploreArea extends Component {
 
   async getOrigin() {
     const { meals } = await getAllOrigin();
-    this.setState({ data: meals, loading: false });
+    this.setState({ data: meals });
   }
 
   allFoods() {
@@ -49,7 +49,6 @@ class ExploreArea extends Component {
 
   renderRecipes() {
     const { meals, type } = this.props;
-    console.log(meals);
     const LENGTH = 12;
     const INITIAL_LENGTH = 0;
     const MAX_LENGTH = meals.length > LENGTH ? LENGTH : meals.length;
@@ -78,18 +77,16 @@ class ExploreArea extends Component {
   }
 
   render() {
-    const { data, loading } = this.state;
+    const { data } = this.state;
     return (
       <div>
         <CustomHeader title="Explorar Origem" />
-        {loading ? (
-          <p>Loadinggg</p>
-        ) : (
+        {
           <CustomDropdown
             data={ data }
             allFoods={ this.allFoods }
           />
-        )}
+        }
         {this.handleRecipes()}
 
         <CustomFooter />
@@ -100,6 +97,7 @@ class ExploreArea extends Component {
 
 const mapStateToProps = (state) => ({
   meals: state.recipesReducer.recipes,
+  isFetching: state.recipesReducer.isFetching,
 });
 const mapDispatchToProps = (dispatch) => ({
   dispatchFoodRecipes: (searchHeader) => dispatch(getFoodRecipes(searchHeader)),
