+++ src/redux/reducer/index.js
@@ -1,3 +1,2 @@
 export { default as loginReducer } from './loginReducer';
-export { default as areaReducer } from './areaReducer';
 export { default as recipesReducer } from './recipesReducer';
