+++ src/services/areaApi.js
@@ -1,19 +1,19 @@
 import {
-  requestArea,
-  requestAreaFailed,
+  failedRequest,
+  requestRecipes,
   selectedArea,
 } from '../redux/actions';
 
 export const getArea = (strArea) => {
   const endpoint = `https://www.themealdb.com/api/json/v1/1/filter.php?a=${strArea}`;
   return async (dispatch) => {
-    dispatch(requestArea());
+    dispatch(requestRecipes());
     try {
       const response = await fetch(endpoint);
       const data = await response.json();
       dispatch(selectedArea(data.meals));
     } catch (error) {
-      dispatch(requestAreaFailed(error));
+      dispatch(failedRequest(error));
     }
   };
 };
