+++ src/store/index.js
@@ -3,18 +3,12 @@ import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import {
   loginReducer,
-  // foodRecipesReducer,
-  // drinkRecipesReducer,
   recipesReducer,
-  areaReducer,
 } from '../redux/reducer';
 
 const rootReducer = combineReducers({
   loginReducer,
-  // foodRecipesReducer,
-  // drinkRecipesReducer,
   recipesReducer,
-  areaReducer,
 });
 
 const store = createStore(
